refactor: remove OptimizingSirToUplcLowering
Use SimpleSirToUplcLowering plus optimizations instead.
Move Inliner to scalus.uplc.transform
