+++ jvm/src/test/scala/scalus/uplc/eval/ExprSizeAndBudgetSpec.scala
@@ -46,9 +46,9 @@ class ExprSizeAndBudgetSpec extends AnyFunSuite {
         assert(encoder.bitSize(uplc) - unitSize - boolSize == 8)
     }
 
-    test("new prelude.List.Cons(true, prelude.List.Nil) size is 103") {
+    test("new prelude.List.Cons(true, prelude.List.Nil) size is 83") {
         val uplc = compile(new prelude.List.Cons(true, prelude.List.Nil)).toUplcOptimized()
-        assert(encoder.bitSize(uplc) == 103)
+        assert(encoder.bitSize(uplc) == 83)
     }
 
     test("prelude.List.cons(true, prelude.List.Nil) size is 123") {
@@ -56,9 +56,9 @@ class ExprSizeAndBudgetSpec extends AnyFunSuite {
         assert(encoder.bitSize(uplc) == 123)
     }
 
-    test("prelude.List.single(true) size is 143") {
+    test("prelude.List.single(true) size is 123") {
         val uplc = compile(prelude.List.single(true)).toUplcOptimized()
-        assert(encoder.bitSize(uplc) == 143)
+        assert(encoder.bitSize(uplc) == 123)
     }
 
     test("equalsInteger(unIData) < equalsData(iData)") {
